requirements.txt
openai
pillar-youtube-upload
dotenv
moviepy
python-dotenv
requests
tempfile